- VisualizeHough Binary/Gradient  * Can select lines in parameter space  * Tool Tips  * other updates - Abstracted HoughBinary can now create an input binary image from edges - Created ControlPanelEdgeThreshold - Created ConfigEdgeThreshold

+++ integration/boofcv-swing/src/main/java/boofcv/gui/StandardAlgConfigPanel.java
@@ -78,7 +78,7 @@ public class StandardAlgConfigPanel extends JPanel implements ActionListener, Ch
 		add(component);
 	}
 
-	protected JConfigLength configLength(ConfigLength initial , int min , int max ) {
+	protected JConfigLength configLength(ConfigLength initial , double min , double max ) {
 		JConfigLength control = new JConfigLength(this,true);
 		control.setValue(initial);
 		control.setLengthBounds(min,max);
@@ -86,7 +86,7 @@ public class StandardAlgConfigPanel extends JPanel implements ActionListener, Ch
 		return control;
 	}
 
-	protected JConfigLength configLength(ConfigLength initial , int min , int max , BoofLambdas.Process listener ) {
+	protected JConfigLength configLength(ConfigLength initial , double min , double max , BoofLambdas.Process listener ) {
 		JConfigLength.Listener cl = (source, fraction, length) -> listener.process();
 
 		JConfigLength control = new JConfigLength(cl,true);
